+++ rotkehlchen/chain/ethereum/modules/balancer/v2/__init__.py
@@ -1,0 +1,2 @@
+from .accountant import Balancerv2Accountant  # noqa: F401
+from .decoder import Balancerv2Decoder  # noqa: F401