+++ rotkehlchen/chain/ethereum/modules/balancer/v2/decoder.py
@@ -4,7 +4,7 @@ from typing import TYPE_CHECKING, Any, Callable, Optional
 from rotkehlchen.accounting.structures.base import HistoryBaseEntry
 from rotkehlchen.accounting.structures.types import HistoryEventSubType, HistoryEventType
 from rotkehlchen.assets.asset import EvmToken
-from rotkehlchen.chain.ethereum.modules.balancer.types import BalancerV1EventTypes
+from rotkehlchen.chain.ethereum.modules.balancer.constants import CPT_BALANCER_V2
 from rotkehlchen.chain.ethereum.utils import asset_normalized_value
 from rotkehlchen.chain.evm.decoding.interfaces import DecoderInterface
 from rotkehlchen.chain.evm.decoding.structures import ActionItem
@@ -16,23 +16,19 @@ from rotkehlchen.logging import RotkehlchenLogsAdapter
 from rotkehlchen.types import ChecksumEvmAddress, EvmTransaction
 from rotkehlchen.utils.misc import hex_or_bytes_to_address, hex_or_bytes_to_int
 
-from .constants import CPT_BALANCER_V1, CPT_BALANCER_V2
-
 if TYPE_CHECKING:
     from rotkehlchen.chain.evm.decoding.base import BaseDecoderTools
     from rotkehlchen.chain.evm.node_inquirer import EvmNodeInquirer
     from rotkehlchen.user_messages import MessagesAggregator
 
 V2_SWAP = b'!p\xc7A\xc4\x151\xae\xc2\x0e|\x10|$\xee\xcf\xdd\x15\xe6\x9c\x9b\xb0\xa8\xdd7\xb1\x84\x0b\x9e\x0b {'  # noqa: E501
-JOIN_V1 = b'c\x98-\xf1\x0e\xfd\x8d\xfa\xaa\xa0\xfc\xc7\xf5\x0b-\x93\xb7\xcb\xa2l\xccH\xad\xee(s"\rH]\xc3\x9a'  # noqa: E501
-EXIT_V1 = b'\xe7L\x91U+d\xc2\xe2\xe7\xbd%V9\xe0\x04\xe6\x93\xbd>\x1d\x01\xcc3\xe6V\x10\xb8j\xfc\xc1\xff\xed'  # noqa: E501
 VAULT_ADDRESS = string_to_evm_address('0xBA12222222228d8Ba445958a75a0704d566BF2C8')
 
 logger = logging.getLogger(__name__)
 log = RotkehlchenLogsAdapter(logger)
 
 
-class BalancerDecoder(DecoderInterface):
+class Balancerv2Decoder(DecoderInterface):
 
     def __init__(
             self,
@@ -47,40 +43,6 @@ class BalancerDecoder(DecoderInterface):
         )
         self.eth = A_ETH.resolve_to_crypto_asset()
         self.weth = A_WETH.resolve_to_evm_token()
-
-    def _decode_v1_pool_event(self, all_logs: list[EvmTxReceiptLog]) -> Optional[list[dict[str, Any]]]:  # noqa: E501
-        """Read the list of logs in search for a Balancer v1 event and return the information
-        needed to decode the transfers made in the transaction to/from the ds proxy
-        """
-        # The transfer event appears after the debt generation event, so we need to transform it
-        target_logs = []
-        for tx_log in all_logs:
-            if tx_log.topics[0] == JOIN_V1 or tx_log.topics[0] == EXIT_V1:
-                target_logs.append(tx_log)
-
-        if len(target_logs) == 0:
-            return None
-
-        events_information = []
-        for target_log in target_logs:
-            token_address = hex_or_bytes_to_address(target_log.topics[2])
-            amount = hex_or_bytes_to_int(target_log.data[0:32])
-            if target_log.topics[0] == JOIN_V1:
-                balancer_event_type = BalancerV1EventTypes.JOIN
-                ds_address = hex_or_bytes_to_address(target_log.topics[1])
-            else:
-                balancer_event_type = BalancerV1EventTypes.EXIT
-                ds_address = target_log.address
-
-            proxy_event_information = {
-                'ds_address': ds_address,
-                'token_address': token_address,
-                'amount': amount,
-                'type': balancer_event_type,
-            }
-            events_information.append(proxy_event_information)
-
-        return events_information
 
     def decode_swap_creation(
             self,
@@ -191,72 +153,6 @@ class BalancerDecoder(DecoderInterface):
 
         return True
 
-    def _maybe_enrich_balancer_v1_events(
-            self,
-            token: 'EvmToken',
-            tx_log: EvmTxReceiptLog,  # pylint: disable=unused-argument
-            transaction: EvmTransaction,  # pylint: disable=unused-argument
-            event: HistoryBaseEntry,
-            action_items: list[ActionItem],  # pylint: disable=unused-argument
-            all_logs: list[EvmTxReceiptLog],
-    ) -> bool:
-        """
-        Enrich balancer v1 transfer to read pool events of:
-        - Depositing in the pool
-        - Withdrawing from the pool
-
-        In balancer v1 pools are managed using a DSProxy so the account doesn't interact
-        directly with the pools.
-        """
-        events_information = self._decode_v1_pool_event(all_logs=all_logs)
-        if events_information is None:
-            return False
-
-        for proxied_event in events_information:
-            if proxied_event['ds_address'] != event.counterparty:
-                continue
-
-            if proxied_event['type'] == BalancerV1EventTypes.JOIN:
-                if (
-                    event.event_type == HistoryEventType.RECEIVE and
-                    event.event_subtype == HistoryEventSubType.NONE
-                ):
-                    event.event_type = HistoryEventType.STAKING
-                    event.event_subtype = HistoryEventSubType.RECEIVE_WRAPPED
-                    event.counterparty = CPT_BALANCER_V1
-                    event.notes = f'Receive {event.balance.amount} {token.symbol} from a Balancer v1 pool'  # noqa: E501
-                    return True
-                if (
-                    event.event_type == HistoryEventType.SPEND and
-                    event.event_subtype == HistoryEventSubType.NONE
-                ):
-                    event.event_type = HistoryEventType.STAKING
-                    event.event_subtype = HistoryEventSubType.DEPOSIT_ASSET
-                    event.counterparty = CPT_BALANCER_V1
-                    event.notes = f'Deposit {event.balance.amount} {token.symbol} to a Balancer v1 pool'  # noqa: E501
-                    return True
-            elif proxied_event['type'] == BalancerV1EventTypes.EXIT:
-                if (
-                    event.event_type == HistoryEventType.RECEIVE and
-                    event.event_subtype == HistoryEventSubType.NONE
-                ):
-                    event.event_type = HistoryEventType.STAKING
-                    event.event_subtype = HistoryEventSubType.REMOVE_ASSET
-                    event.counterparty = CPT_BALANCER_V1
-                    event.notes = f'Receive {event.balance.amount} {token.symbol} after removing liquidity from a Balancer v1 pool'  # noqa: E501
-                    return True
-                if (
-                    event.event_type == HistoryEventType.SPEND and
-                    event.event_subtype == HistoryEventSubType.NONE
-                ):
-                    event.event_type = HistoryEventType.STAKING
-                    event.event_subtype = HistoryEventSubType.RETURN_WRAPPED
-                    event.counterparty = CPT_BALANCER_V1
-                    event.notes = f'Return {event.balance.amount} {token.symbol} to a Balancer v1 pool'  # noqa: E501
-                    return True
-
-        return False
-
     # -- DecoderInterface methods
 
     def addresses_to_decoders(self) -> dict[ChecksumEvmAddress, tuple[Any, ...]]:
@@ -267,8 +163,7 @@ class BalancerDecoder(DecoderInterface):
     def enricher_rules(self) -> list[Callable]:
         return [
             self._maybe_enrich_balancer_v2_transfers,
-            self._maybe_enrich_balancer_v1_events,
         ]
 
     def counterparties(self) -> list[str]:
-        return [CPT_BALANCER_V1, CPT_BALANCER_V2]
+        return [CPT_BALANCER_V2]