+++ rotkehlchen/db/reports.py
@@ -100,6 +100,7 @@ class DBAccountingReports():
                     (report_id, 'account_for_assets_movements', 'bool', settings.account_for_assets_movements),  # noqa: E501
                     (report_id, 'cost_basis_method', 'string', settings.cost_basis_method.serialize()),  # noqa: E501
                     (report_id, 'eth_staking_taxable_after_withdrawal_enabled', 'bool', settings.eth_staking_taxable_after_withdrawal_enabled),  # noqa: E501
+                    (report_id, 'include_fees_in_cost_basis', 'bool', settings.include_fees_in_cost_basis),  # noqa: E501
                 ])
 
         return report_id