+++ rotkehlchen/tests/conftest.py
@@ -35,6 +35,8 @@ SUBPROCESS_TIMEOUT = 30
 
 
 class TestEnvironment(SerializableEnumMixin):
+    __test__ = False  # tell pytest not to collect this class
+
     STANDARD = auto()  # test during normal development
     NIGHTLY = auto()  # all tests
     NFTS = auto()  # nft tests
@@ -108,6 +110,7 @@ if sys.platform == 'darwin':
         created as a sub directory of the base temporary
         directory.  The returned object is a `py.path.local`_
         path object.
+        # pytest-deadfixtures ignore
         """
         name = request.node.name
         name = re.sub(r'[\W]', '_', name)